1merge_init_commit2d_list_refactor

+++ get_pos_val.c
@@ -5,7 +5,7 @@ size_t	get_pos(t_dlist *stack_elem)
 	return (((t_ps_data *)(stack_elem->content))->pos);
 }
 
-int		get_flag(t_dlist *stack_elem)
+size_t  get_flag(t_dlist *stack_elem)
 {
 	return (((t_ps_data *)(stack_elem->content))->flag);
 }
@@ -15,7 +15,7 @@ size_t	*set_flag(t_dlist *stack_elem)
 	return &(((t_ps_data *)(stack_elem->content))->flag);
 }
 
-int	seek_pos(t_dlist *stack, size_t part, size_t size)
+int 	seek_pos(t_dlist *stack, size_t part, size_t size)
 {
 	size_t	pos;
